refactor(app/errors): cover base documentation + refactor
#15
